lint: [move-only] Move util functions to util.rs
